Fixes connection issues (maybe)
Some people were experiencing issues with all games not being found, I myself can't seem to replicate this issue tho, so I don't know for sure if this fixes it. If not, please just wait for the rewrite (vue-rewrite branch) as it will be infinitely better anyways. Thanks for the continued support, could not have done it without you guys.