0.7.0 Answer Correct, Answer Explorer & Leaderboard Flood
Added Answer explorer to view answers of games
Added Answer correct button to automatically send the answer for the current question (this will be automated soon)
Added Leaderboard flood, this hack fills the leaderboard with one specified name, this works best with a playergroup with one player.
Added Leaderboard flood with auto answer, nobody will be able to answer, and you will get all the points!